+++ src/main/java/com/mindhub/homebanking/configurations/WebSecurityConfiguration.java
@@ -1,17 +1,14 @@
 package com.mindhub.homebanking.configurations;
 import com.mindhub.homebanking.models.Client;
 import com.mindhub.homebanking.repositories.ClientRepository;
-import com.mindhub.homebanking.services.NotificationService;
+import com.mindhub.homebanking.services.NotificationServiceImp;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.authentication.configuration.GlobalAuthenticationConfigurerAdapter;
 import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
 
 import javax.servlet.http.HttpSession;
 import java.time.LocalTime;
@@ -28,7 +25,7 @@ public class WebSecurityConfiguration extends GlobalAuthenticationConfigurerAdap
     HttpSession session;
 
     @Autowired
-    NotificationService notificationService;
+    NotificationServiceImp notificationService;
 
 
     @Override
