+++ src/main/java/com/mindhub/homebanking/controller/CardController.java
@@ -4,7 +4,9 @@ import com.mindhub.homebanking.models.CardColor;
 import com.mindhub.homebanking.models.CardType;
 import com.mindhub.homebanking.models.Client;
 import com.mindhub.homebanking.services.CardService;
+import com.mindhub.homebanking.services.CardServiceImp;
 import com.mindhub.homebanking.services.ClientService;
+import com.mindhub.homebanking.services.ClientServiceImp;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.context.i18n.LocaleContextHolder;
@@ -12,7 +14,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
-import javax.servlet.http.HttpSession;
+
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
