+++ src/main/java/com/mindhub/homebanking/controller/SharedTransactionController.java
@@ -2,10 +2,7 @@ package com.mindhub.homebanking.controller;
 
 import com.mindhub.homebanking.dtos.SharedTransactionDTO;
 import com.mindhub.homebanking.models.Client;
-import com.mindhub.homebanking.models.SharedTransaction;
-import com.mindhub.homebanking.services.ClientService;
-import com.mindhub.homebanking.services.SharedTransactionService;
-import com.mindhub.homebanking.services.TransactionService;
+import com.mindhub.homebanking.services.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.context.i18n.LocaleContextHolder;
@@ -14,9 +11,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
 
 @RestController
 @RequestMapping("/api")
@@ -34,6 +29,8 @@ public class SharedTransactionController {
     @Autowired
     private MessageSource mensajes;
 
+    public SharedTransactionController() {
+    }
 
 
     @PostMapping("/transactions/shared")
