+++ src/main/java/com/mindhub/homebanking/services/SharedTransactionService.java
@@ -1,98 +1,15 @@
 package com.mindhub.homebanking.services;
 
 import com.mindhub.homebanking.dtos.SharedTransactionDTO;
-import com.mindhub.homebanking.models.Account;
 import com.mindhub.homebanking.models.Client;
-import com.mindhub.homebanking.models.SharedTransaction;
-import com.mindhub.homebanking.models.SharedTransactionAccount;
-import com.mindhub.homebanking.repositories.AccountRepository;
-import com.mindhub.homebanking.repositories.SharedTransactionAccountRepository;
-import com.mindhub.homebanking.repositories.SharedTransactionRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
-import java.awt.Toolkit;
-import java.awt.datatransfer.Clipboard;
-import java.awt.datatransfer.StringSelection;
-import java.util.UUID;
+public interface SharedTransactionService {
 
-@Service
-public class SharedTransactionService {
+    public SharedTransactionDTO findById(long id);
 
-    @Autowired
-    private SharedTransactionRepository sharedTransactionRepository;
+    public String makeSharedTransaction(Long amount, int sharedAccounts,String fromAccountNumber, String toAccountNumber, String description, Client clientelogueado);
 
-    @Autowired
-    private SharedTransactionAccountRepository sharedTransactionAccountRepository;
-    @Autowired
-    private AccountRepository accountRepository;
+    public boolean paySharedTransaction(String tokenId, String fromAccountNumber, Client clientelogueado);
 
-    @Autowired
-    private AccountService accountService;
-
-    @Autowired
-    private NotificationService notificationService;
-
-    @Autowired
-    private TransactionService transactionService;
-
-
-
-    public SharedTransactionDTO findById(long id){
-        return  new SharedTransactionDTO(sharedTransactionRepository.findById(id).orElse(null));
-    }
-
-
-    public String makeSharedTransaction(Long amount, int sharedAccounts,String fromAccountNumber, String toAccountNumber, String description, Client clientelogueado){
-        long amountPerAccount = (long) (amount /  sharedAccounts);
-        String tokenId = (UUID.randomUUID()).toString().toUpperCase();
-        Account account = accountRepository.findByNumber(toAccountNumber).orElse(null);
-        SharedTransaction sharedTransaction = new SharedTransaction(account,amount,amountPerAccount,sharedAccounts,tokenId,description);
-        sharedTransaction = sharedTransactionRepository.save(sharedTransaction);
-        accountRepository.save(account);
-
-        System.out.println("Token: " + tokenId);
-        String linkPago= "http://localhost:8080/web/pay-shared-transfers.html?tokenId=" + tokenId;
-
-        if(clientelogueado.isHasTelegram()){
-            notificationService.sendNotification(clientelogueado.getEmail(),
-                    "Se ha realizado un pago compartido desde " +  toAccountNumber +
-                            " hacia la cuenta " + fromAccountNumber + " con monto " + amount +
-                    ". Link de pago: " + linkPago);
-        }
-
-        return linkPago;
-    }
-
-        public boolean paySharedTransaction(String tokenId, String fromAccountNumber, Client clientelogueado){
-
-        Account account = accountRepository.findByNumber(fromAccountNumber).orElse(null);
-        SharedTransaction sharedTransaction = sharedTransactionRepository.findByTokenId(tokenId).orElse(null);
-
-        transactionService.makeTransaction(sharedTransaction.getTotalAmount(), ("Pago compartido " + sharedTransaction.getId()),fromAccountNumber,sharedTransaction.getAccount().getNumber(),clientelogueado);
-
-
-        if(account != null && sharedTransaction != null){
-            SharedTransactionAccount sharedTransactionAccount = new SharedTransactionAccount(account,sharedTransaction);
-            account.addPagoSharedTransaction(sharedTransactionAccount);
-            sharedTransaction.addToAccount(sharedTransactionAccount);
-            accountRepository.save(account);
-            sharedTransactionRepository.save(sharedTransaction);
-            sharedTransactionAccountRepository.save(sharedTransactionAccount);
-
-            if(clientelogueado.isHasTelegram()){
-                notificationService.sendNotification(clientelogueado.getEmail(),
-                        "Se ha pagado un pago compartido desde " + fromAccountNumber +
-                                " hacia la cuenta " + sharedTransaction.getAccount() + " con monto " + sharedTransaction.getParcialAmount());
-            }
-
-            return true;
-        }else {
-            return false;
-        }
-    }
-
-    public SharedTransactionDTO findByTokenId(String tokenId){
-        return  new SharedTransactionDTO(sharedTransactionRepository.findByTokenId(tokenId).orElse(null));
-    }
+    public SharedTransactionDTO findByTokenId(String tokenId);
 }
