+++ src/test/java/com/mindhub/homebanking/NotificationServiceTest.java
@@ -1,0 +1,8 @@
+package com.mindhub.homebanking;
+
+import static org.mockito.Mockito.mock;
+
+
+public class NotificationServiceTest {
+
+}
